agents/requirements.txt
strands-agents
bedrock-agentcore
aws-opentelemetry-distro>=0.10.1
boto3
